+++ src/lib/include/heap.h
@@ -2,5 +2,5 @@
 #define __HEAP__H__
 #include "stddef.h"
 #include "string.h"
-void *malloc(size_t n);
+void *malloc(unsigned long n);
 #endif
